All role & permission set and Api create for Uttara Motors domain and QrCode print update

+++ app/Http/Controllers/Backend/PermissionController.php
@@ -12,21 +12,19 @@ class PermissionController extends Controller
 {
     public function index(Request $request)
     {
-        if (!Auth::user()->can('permission.manage')) {
-            return abort(403, 'Unauthorized');
-        }
+        abort_unless(Auth::user()->can('permission.manage'), 403, 'Unauthorized');
 
         try {
             $permission_groups = User::getpermissionGroups();
             $all_permissions = Permission::query()->orderBy('created_at', 'desc');
 
             $all_permissions = $all_permissions->get();
-            if($request->ajax()){
+            if ($request->ajax()) {
                 $data = view('backend.pages.permission.showPermissionList', compact('all_permissions', 'permission_groups'))->render();
                 return response()->json(['data' => $data]);
             }
 
-            return view('backend.pages.permission.index',compact('all_permissions', 'permission_groups'));
+            return view('backend.pages.permission.index', compact('all_permissions', 'permission_groups'));
         } catch (\Exception $th) {
             return response()->json([
                 'atert_type' => 'error',
@@ -37,18 +35,14 @@ class PermissionController extends Controller
 
     public function create()
     {
-        if (!Auth::user()->can('permission.create')) {
-            return abort(403, 'Unauthorized');
-        }
+        abort_unless(Auth::user()->can('permission.create'), 403, 'Unauthorized');
         return view('backend.pages.permission.create');
     }
 
 
     public function store(Request $request)
     {
-        if (!Auth::user()->can('permission.create')) {
-            return abort(403, 'Unauthorized');
-        }
+        abort_unless(Auth::user()->can('permission.create'), 403, 'Unauthorized');
         // Validation Data
         // $validate = $request->validate(
         //     [
@@ -62,10 +56,10 @@ class PermissionController extends Controller
         //         //'name.max'            => "Permission Name must not be greater than 50 characters",
         //     ]
         // );
-        $countPermission = count($request->name)-1;
-         //dd($countPermission);
-        if ( $countPermission  !== 0){
-            for ( $i=0; $i <$countPermission; $i++ ){
+        $countPermission = count($request->name) - 1;
+        //dd($countPermission);
+        if ($countPermission  !== 0) {
+            for ($i = 0; $i < $countPermission; $i++) {
                 $permission_new                 = new Permission();
                 $permission_new->group_name     = $request->group_name;
                 $permission_new->guard_name     = 'web';
@@ -73,17 +67,15 @@ class PermissionController extends Controller
                 $permission_new->save();
             }
             return redirect()->route('dashboard.permission.index')->with('Permission Created Successfully');
-        }
-        else {
+        } else {
             dd('error notification');
         }
     }
 
     public function edit($group_name)
     {
-        if (!Auth::user()->can('permission.edit')) {
-            return abort(403, 'Unauthorized');
-        }
+        abort_unless(Auth::user()->can('permission.edit'), 403, 'Unauthorized');
+
         $all_permission = Permission::select('name')->where('group_name', $group_name)->get();
         //dd($all_permission);
         return view('backend.pages.permission.edit', compact('all_permission', 'group_name'));
@@ -91,26 +83,26 @@ class PermissionController extends Controller
 
     public function update(Request $request, $group_name)
     {
+        abort_unless(Auth::user()->can('permission.edit'), 403, 'Unauthorized');
+
         $permission = Permission::select('group_name')->where('group_name', $group_name)->groupBy('group_name')->get();
 
-        if ( $permission[0]['group_name'] == $group_name){
+        if ($permission[0]['group_name'] == $group_name) {
             //dd($group_name);
-            $countPermission = count($request->name)-1;
+            $countPermission = count($request->name) - 1;
             //dd($countPermission);
 
             // if ( $countPermission != NULL ){
-                Permission::where('group_name', $group_name)->delete();
-                for ( $i=0; $i < $countPermission; $i++ ){
-                    $permission_new                 = new Permission();
+            Permission::where('group_name', $group_name)->delete();
+            for ($i = 0; $i < $countPermission; $i++) {
+                $permission_new                 = new Permission();
 
-                    $permission_new->group_name     = $request->group_name;
-                    $permission_new->name           = $request->name[$i];
-                    $permission_new->guard_name     = 'web';
-                    $permission_new->save();
-
-
-                }
-                return redirect()->route('dashboard.permission.index')->with('Permission Updated Successfullt');
+                $permission_new->group_name     = $request->group_name;
+                $permission_new->name           = $request->name[$i];
+                $permission_new->guard_name     = 'web';
+                $permission_new->save();
+            }
+            return redirect()->route('dashboard.permission.index')->with('Permission Updated Successfullt');
             //}
             //dd($permission);
         }
@@ -118,28 +110,23 @@ class PermissionController extends Controller
 
     public function destroy(Request $request, $group_name)
     {
-        if (!Auth::user()->can('permission.delete')) {
-            return abort(403, 'Unauthorized');
-        }
+        abort_unless(Auth::user()->can('permission.delete'), 403, 'Unauthorized');
 
         $permissions  = Permission::where('group_name', $group_name)->get();
         //dd($permissions);
-        if (!is_null($permissions) && $permissions->isNotEmpty())
-        {
-        // Loop through each permission and delete it along with related roles
-        foreach ($permissions as $permission) {
-            $permission->roles()->detach(); // Remove related roles
-            $permission->delete(); // Delete the permission
-        }
+        if (!is_null($permissions) && $permissions->isNotEmpty()) {
+            // Loop through each permission and delete it along with related roles
+            foreach ($permissions as $permission) {
+                $permission->roles()->detach(); // Remove related roles
+                $permission->delete(); // Delete the permission
+            }
             return response()->json([
                 'success' => true,
                 'status' => 'warning',
                 "message" => "Permissions and related roles deleted successfully.",
                 "url" => route('dashboard.permission.index')
             ]);
-        }
-
-        else {
+        } else {
             return response()->json([
                 'status' => 'warning',
                 "message" => "No permissions found for the specified group name.",
